requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
Django==3.0.3
django-flatpickr==1.0.1
django-tinymce==2.8.0
django-zxcvbn-password==2.1.0
image==1.5.28
Pillow==7.0.0
pytz==2019.3
soupsieve==1.9.5
sqlparse==0.3.0
zxcvbn==4.4.28
